+++ emain/emain.ts
@@ -10,8 +10,6 @@ import * as path from "path";
 import { PNG } from "pngjs";
 import { sprintf } from "sprintf-js";
 import { Readable } from "stream";
-import * as util from "util";
-import winston from "winston";
 import * as services from "../frontend/app/store/services";
 import { initElectronWshrpc, shutdownWshrpc } from "../frontend/app/store/wshrpcutil";
 import { getWebServerEndpoint } from "../frontend/util/endpoints";
@@ -25,7 +23,6 @@ import {
     getGlobalIsRelaunching,
     setForceQuit,
     setGlobalIsQuitting,
-    setGlobalIsRelaunching,
     setGlobalIsStarting,
     setWasActive,
     setWasInFg,
@@ -35,16 +32,19 @@ import { handleCtrlShiftState } from "./emain-util";
 import { getIsWaveSrvDead, getWaveSrvProc, getWaveSrvReady, getWaveVersion, runWaveSrv } from "./emain-wavesrv";
 import {
     createBrowserWindow,
+    createNewWaveWindow,
     focusedWaveWindow,
     getAllWaveWindows,
     getWaveWindowById,
     getWaveWindowByWebContentsId,
     getWaveWindowByWorkspaceId,
+    relaunchBrowserWindows,
     WaveBrowserWindow,
 } from "./emain-window";
 import { ElectronWshClient, initElectronWshClient } from "./emain-wsh";
 import { getLaunchSettings } from "./launchsettings";
-import { getAppMenu } from "./menu";
+import { log } from "./log";
+import { instantiateAppMenu, makeAppMenu } from "./menu";
 import {
     getElectronAppBasePath,
     getElectronAppUnpackedBasePath,
@@ -65,30 +65,7 @@ electron.nativeTheme.themeSource = "dark";
 
 let webviewFocusId: number = null; // set to the getWebContentsId of the webview that has focus (null if not focused)
 let webviewKeys: string[] = []; // the keys to trap when webview has focus
-const oldConsoleLog = console.log;
 
-const loggerTransports: winston.transport[] = [
-    new winston.transports.File({ filename: path.join(waveDataDir, "waveapp.log"), level: "info" }),
-];
-if (isDev) {
-    loggerTransports.push(new winston.transports.Console());
-}
-const loggerConfig = {
-    level: "info",
-    format: winston.format.combine(
-        winston.format.timestamp({ format: "YYYY-MM-DD HH:mm:ss.SSS" }),
-        winston.format.printf((info) => `${info.timestamp} ${info.message}`)
-    ),
-    transports: loggerTransports,
-};
-const logger = winston.createLogger(loggerConfig);
-function log(...msg: any[]) {
-    try {
-        logger.info(util.format(...msg));
-    } catch (e) {
-        oldConsoleLog(...msg);
-    }
-}
 console.log = log;
 console.log(
     sprintf(
@@ -375,34 +352,6 @@ electron.ipcMain.on("open-native-path", (event, filePath: string) => {
     );
 });
 
-async function createNewWaveWindow(): Promise<void> {
-    log("createNewWaveWindow");
-    const clientData = await services.ClientService.GetClientData();
-    const fullConfig = await services.FileService.GetFullConfig();
-    let recreatedWindow = false;
-    const allWindows = getAllWaveWindows();
-    if (allWindows.length === 0 && clientData?.windowids?.length >= 1) {
-        console.log("no windows, but clientData has windowids, recreating first window");
-        // reopen the first window
-        const existingWindowId = clientData.windowids[0];
-        const existingWindowData = (await services.ObjectService.GetObject("window:" + existingWindowId)) as WaveWindow;
-        if (existingWindowData != null) {
-            const win = await createBrowserWindow(existingWindowData, fullConfig, { unamePlatform });
-            await win.waveReadyPromise;
-            win.show();
-            recreatedWindow = true;
-        }
-    }
-    if (recreatedWindow) {
-        console.log("recreated window, returning");
-        return;
-    }
-    console.log("creating new window");
-    const newBrowserWindow = await createBrowserWindow(null, fullConfig, { unamePlatform });
-    await newBrowserWindow.waveReadyPromise;
-    newBrowserWindow.show();
-}
-
 electron.ipcMain.on("set-window-init-status", (event, status: "ready" | "wave-ready") => {
     const tabView = getWaveTabViewByWebContentsId(event.sender.id);
     if (tabView == null || tabView.initResolve == null) {
@@ -481,10 +430,10 @@ electron.ipcMain.on("contextmenu-show", (event, menuDefArr?: ElectronContextMenu
     if (menuDefArr?.length === 0) {
         return;
     }
-    const menu = menuDefArr ? convertMenuDefArrToMenu(menuDefArr) : instantiateAppMenu();
-    // const { x, y } = electron.screen.getCursorScreenPoint();
-    // const windowPos = window.getPosition();
-    menu.popup();
+    fireAndForget(async () => {
+        const menu = menuDefArr ? convertMenuDefArrToMenu(menuDefArr) : await instantiateAppMenu();
+        menu.popup();
+    });
     event.returnValue = true;
 });
 
@@ -559,18 +508,6 @@ function convertMenuDefArrToMenu(menuDefArr: ElectronContextMenuItem[]): electro
         menuItems.push(menuItem);
     }
     return electron.Menu.buildFromTemplate(menuItems);
-}
-
-function instantiateAppMenu(): electron.Menu {
-    return getAppMenu({
-        createNewWaveWindow,
-        relaunchBrowserWindows,
-    });
-}
-
-function makeAppMenu() {
-    const menu = instantiateAppMenu();
-    electron.Menu.setApplicationMenu(menu);
 }
 
 function hideWindowWithCatch(window: WaveBrowserWindow) {
@@ -649,37 +586,6 @@ process.on("uncaughtException", (error) => {
     electronApp.quit();
 });
 
-async function relaunchBrowserWindows(): Promise<void> {
-    console.log("relaunchBrowserWindows");
-    setGlobalIsRelaunching(true);
-    const windows = getAllWaveWindows();
-    for (const window of windows) {
-        console.log("relaunch -- closing window", window.waveWindowId);
-        window.close();
-    }
-    setGlobalIsRelaunching(false);
-
-    const clientData = await services.ClientService.GetClientData();
-    const fullConfig = await services.FileService.GetFullConfig();
-    const wins: WaveBrowserWindow[] = [];
-    for (const windowId of clientData.windowids.slice().reverse()) {
-        const windowData: WaveWindow = await services.WindowService.GetWindow(windowId);
-        if (windowData == null) {
-            console.log("relaunch -- window data not found, closing window", windowId);
-            await services.WindowService.CloseWindow(windowId, true);
-            continue;
-        }
-        console.log("relaunch -- creating window", windowId, windowData);
-        const win = await createBrowserWindow(windowData, fullConfig, { unamePlatform });
-        wins.push(win);
-    }
-    for (const win of wins) {
-        await win.waveReadyPromise;
-        console.log("show window", win.waveWindowId);
-        win.show();
-    }
-}
-
 async function appMain() {
     // Set disableHardwareAcceleration as early as possible, if required.
     const launchSettings = getLaunchSettings();
@@ -694,7 +600,6 @@ async function appMain() {
         electronApp.quit();
         return;
     }
-    makeAppMenu();
     try {
         await runWaveSrv(handleWSEvent);
     } catch (e) {
@@ -715,6 +620,7 @@ async function appMain() {
     } catch (e) {
         console.log("error initializing wshrpc", e);
     }
+    makeAppMenu();
     await configureAutoUpdater();
     setGlobalIsStarting(false);
     if (fullConfig?.settings?.["window:maxtabcachesize"] != null) {
